+++ src/pake/mod.rs
@@ -10,19 +10,19 @@ mod cpace;
 #[cfg(test)]
 mod tests;
 
+use crate::EncodedSizeUser;
 use crate::Input;
 use core::fmt::Debug;
 pub use cpace::CPaceRistretto255;
-use ml_kem::EncodedSizeUser;
-use rand_core::{CryptoRng, RngCore};
+use rand_core::CryptoRng;
 
 pub trait Pake: EncodedSizeUser {
     type InitMessage: EncodedSizeUser + Debug + PartialEq;
     type RespondMessage: EncodedSizeUser + Debug + PartialEq;
     type Output: EncodedSizeUser + Debug + PartialEq;
 
-    fn init<R: RngCore + CryptoRng>(input: &Input, rng: &mut R) -> (Self::InitMessage, Self);
-    fn respond<R: RngCore + CryptoRng>(
+    fn init<R: CryptoRng>(input: &Input, rng: &mut R) -> (Self::InitMessage, Self);
+    fn respond<R: CryptoRng>(
         input: &Input,
         init_message: &Self::InitMessage,
         rng: &mut R,