+++ src/protocol.rs
@@ -12,21 +12,21 @@ use crate::errors::PakeKemError;
 use crate::messages::{MessageOne, MessageThree, MessageTwo};
 use crate::pake::CPaceRistretto255;
 use crate::pake::Pake;
+use crate::{Encoded, EncodedSizeUser};
 use core::ops::{Add, Sub};
 use hkdf::hmac::digest::array::typenum::U32;
 use hkdf::hmac::digest::array::Array;
-use hkdf::hmac::digest::core_api::BlockSizeUser;
+use hkdf::hmac::digest::block_api::BlockSizeUser;
 use hkdf::hmac::digest::typenum::Sum;
 use hkdf::hmac::digest::FixedOutput;
 use hkdf::hmac::digest::OutputSizeUser;
 use hkdf::hmac::{EagerHash, Hmac, KeyInit, Mac};
 use hkdf::HkdfExtract;
-use kem::{Decapsulate, Encapsulate};
-use ml_kem::ArraySize;
-use ml_kem::Encoded;
-use ml_kem::EncodedSizeUser;
-use ml_kem::KemCore;
-use rand_core::{CryptoRng, RngCore};
+use ml_kem::array::ArraySize;
+use ml_kem::kem::KeySizeUser;
+use ml_kem::kem::TryDecapsulate;
+use ml_kem::{Encapsulate, Kem, KeyExport, TryKeyInit};
+use rand_core::CryptoRng;
 
 type Result<T> = core::result::Result<T, PakeKemError>;
 
@@ -38,7 +38,7 @@ pub trait CipherSuite {
     /// The PAKE protocol to use
     type Pake: Pake;
     /// The key encapsulation mechanism to use
-    type Kem: KemCore;
+    type Kem: Kem;
     /// The hashing function to use
     type Hash: EagerHash + FixedOutput;
 }
@@ -72,7 +72,9 @@ impl Input {
 
 /// The output of the pake-kem protocol
 #[derive(Clone, Debug, PartialEq, Eq)]
-pub struct Output<CS: CipherSuite>(pub Array<u8, <CS::Hash as OutputSizeUser>::OutputSize>);
+pub struct Output<CS: CipherSuite>(
+    pub Array<u8, <<CS::Hash as EagerHash>::Core as OutputSizeUser>::OutputSize>,
+);
 
 /// The main struct for the initiator of the pake-kem protocol
 #[derive(Debug)]
@@ -84,10 +86,7 @@ where
         Sub<<<CS::Hash as EagerHash>::Core as BlockSizeUser>::BlockSize, Output = U32>,
 {
     /// The first step of pake-kem, where the initiator starts the protocol
-    pub fn start<R: RngCore + CryptoRng>(
-        input: &Input,
-        rng: &mut R,
-    ) -> Result<(Self, MessageOne<CS>)> {
+    pub fn start<R: CryptoRng>(input: &Input, rng: &mut R) -> Result<(Self, MessageOne<CS>)> {
         let (init_message, state) = CS::Pake::init(input, rng);
 
         Ok((Self(state), MessageOne { init_message }))
@@ -95,7 +94,7 @@ where
 
     /// The third step of pake-kem, where the initiator finishes its role in the protocol, with
     /// input from the second message created by the responder
-    pub fn finish<R: RngCore + CryptoRng>(
+    pub fn finish<R: CryptoRng>(
         self,
         message_two: &MessageTwo<CS>,
         rng: &mut R,
@@ -107,24 +106,21 @@ where
 
         // First, check the mac on ek
         let mut mac_verifier = Hmac::<CS::Hash>::new(&mac_key);
-        mac_verifier.update(&message_two.ek.as_bytes());
+        mac_verifier.update(&message_two.ek.to_bytes());
         mac_verifier.verify_slice(&message_two.ek_tag)?;
 
         // Encapsulate a shared key to the holder of the decapsulation key, receive the shared
         // secret `k_send` and the encapsulated form `ct`.
-        let (ct, k_send) = message_two
-            .ek
-            .encapsulate(rng)
-            .map_err(|_| PakeKemError::Deserialization)?;
+        let (ct, k_send) = message_two.ek.encapsulate_with_rng(rng);
 
         // Next, construct another mac
         let mut mac_builder = Hmac::<CS::Hash>::new(&mac_key);
-        mac_builder.update(&message_two.ek.as_bytes());
+        mac_builder.update(&message_two.ek.to_bytes());
         mac_builder.update(ct.as_slice());
         let mac = mac_builder.finalize().into_bytes();
 
         let mut hkdf = HkdfExtract::<CS::Hash>::new(None);
-        hkdf.input_ikm(&message_two.ek.as_bytes());
+        hkdf.input_ikm(&message_two.ek.to_bytes());
         hkdf.input_ikm(ct.as_slice());
         hkdf.input_ikm(&mac_key);
         hkdf.input_ikm(&session_key);
@@ -136,11 +132,18 @@ where
 }
 
 /// The main struct for the responder of the pake-kem protocol
-#[derive(Debug)]
 pub struct Responder<CS: CipherSuite> {
     pake_output: <CS::Pake as Pake>::Output,
-    dk: <CS::Kem as KemCore>::DecapsulationKey,
-    ek: <CS::Kem as KemCore>::EncapsulationKey,
+    dk: <CS::Kem as Kem>::DecapsulationKey,
+    ek: <CS::Kem as Kem>::EncapsulationKey,
+}
+
+impl<CS: CipherSuite> core::fmt::Debug for Responder<CS> {
+    fn fmt(&self, f: &mut core::fmt::Formatter<'_>) -> core::fmt::Result {
+        f.debug_struct("Responder")
+            .field("pake_output", &self.pake_output)
+            .finish_non_exhaustive()
+    }
 }
 
 impl<CS: CipherSuite> Responder<CS>
@@ -150,7 +153,7 @@ where
 {
     /// The second step of pake-kem, where the responder starts its role in the protocol
     /// with input from the first message created by the initiator
-    pub fn start<R: RngCore + CryptoRng>(
+    pub fn start<R: CryptoRng>(
         input: &Input,
         message_one: &MessageOne<CS>,
         rng: &mut R,
@@ -163,10 +166,11 @@ where
         };
         let (mac_key, _) = pake_output_into_keys::<CS>(pake_output.as_bytes());
 
-        let (decapsulation_key, encapsulation_key) = CS::Kem::generate(rng);
+        let (decapsulation_key, encapsulation_key) = CS::Kem::generate_keypair_from_rng(rng);
 
-        let ek_bytes = encapsulation_key.as_bytes();
-        let ek_cloned = <CS::Kem as KemCore>::EncapsulationKey::from_bytes(&ek_bytes);
+        let ek_bytes = encapsulation_key.to_bytes();
+        let ek_cloned = <CS::Kem as Kem>::EncapsulationKey::new(&ek_bytes)
+            .map_err(|_| PakeKemError::Deserialization)?;
 
         let mut mac_builder = Hmac::<CS::Hash>::new(&mac_key);
         mac_builder.update(&ek_bytes);
@@ -192,17 +196,17 @@ where
         let (mac_key, session_key) = pake_output_into_keys::<CS>(self.pake_output.as_bytes());
 
         let mut mac_verifier = Hmac::<CS::Hash>::new(&mac_key);
-        mac_verifier.update(&self.ek.as_bytes());
+        mac_verifier.update(&self.ek.to_bytes());
         mac_verifier.update(message_three.ct.as_slice());
         mac_verifier.verify_slice(&message_three.ct_tag)?;
 
         let k_recv = self
             .dk
-            .decapsulate(&message_three.ct)
+            .try_decapsulate(&message_three.ct)
             .map_err(|_| PakeKemError::Deserialization)?;
 
         let mut hkdf = HkdfExtract::<CS::Hash>::new(None);
-        hkdf.input_ikm(&self.ek.as_bytes());
+        hkdf.input_ikm(&self.ek.to_bytes());
         hkdf.input_ikm(message_three.ct.as_slice());
         hkdf.input_ikm(&mac_key);
         hkdf.input_ikm(&session_key);
@@ -230,8 +234,8 @@ where
 impl<CS: CipherSuite> EncodedSizeUser for Initiator<CS> {
     type EncodedSize = <CS::Pake as EncodedSizeUser>::EncodedSize;
 
-    fn from_bytes(enc: &Encoded<Self>) -> Self {
-        Self(CS::Pake::from_bytes(enc))
+    fn from_bytes(enc: &Encoded<Self>) -> core::result::Result<Self, PakeKemError> {
+        Ok(Self(CS::Pake::from_bytes(enc)?))
     }
 
     fn as_bytes(&self) -> Encoded<Self> {
@@ -241,56 +245,59 @@ impl<CS: CipherSuite> EncodedSizeUser for Initiator<CS> {
 
 impl<CS: CipherSuite> EncodedSizeUser for Responder<CS>
 where
+    CS::Kem: ml_kem::kem::FromSeed,
+    <CS::Kem as Kem>::DecapsulationKey: KeyExport + KeyInit,
+    <CS::Kem as Kem>::EncapsulationKey: KeyExport + TryKeyInit,
     // Concatenation clauses
-    <<CS::Kem as KemCore>::DecapsulationKey as EncodedSizeUser>::EncodedSize:
-        Add<<<CS::Kem as KemCore>::EncapsulationKey as EncodedSizeUser>::EncodedSize>,
+    <<CS::Kem as Kem>::DecapsulationKey as KeySizeUser>::KeySize:
+        Add<<<CS::Kem as Kem>::EncapsulationKey as KeySizeUser>::KeySize>,
     Sum<
-        <<CS::Kem as KemCore>::DecapsulationKey as EncodedSizeUser>::EncodedSize,
-        <<CS::Kem as KemCore>::EncapsulationKey as EncodedSizeUser>::EncodedSize,
+        <<CS::Kem as Kem>::DecapsulationKey as KeySizeUser>::KeySize,
+        <<CS::Kem as Kem>::EncapsulationKey as KeySizeUser>::KeySize,
     >: ArraySize
         + Add<<<CS::Pake as Pake>::Output as EncodedSizeUser>::EncodedSize>
         + Sub<
-            <<CS::Kem as KemCore>::DecapsulationKey as EncodedSizeUser>::EncodedSize,
-            Output = <<CS::Kem as KemCore>::EncapsulationKey as EncodedSizeUser>::EncodedSize,
+            <<CS::Kem as Kem>::DecapsulationKey as KeySizeUser>::KeySize,
+            Output = <<CS::Kem as Kem>::EncapsulationKey as KeySizeUser>::KeySize,
         >,
     Sum<
         Sum<
-            <<CS::Kem as KemCore>::DecapsulationKey as EncodedSizeUser>::EncodedSize,
-            <<CS::Kem as KemCore>::EncapsulationKey as EncodedSizeUser>::EncodedSize,
+            <<CS::Kem as Kem>::DecapsulationKey as KeySizeUser>::KeySize,
+            <<CS::Kem as Kem>::EncapsulationKey as KeySizeUser>::KeySize,
         >,
         <<CS::Pake as Pake>::Output as EncodedSizeUser>::EncodedSize,
     >: ArraySize
         + Sub<
             Sum<
-                <<CS::Kem as KemCore>::DecapsulationKey as EncodedSizeUser>::EncodedSize,
-                <<CS::Kem as KemCore>::EncapsulationKey as EncodedSizeUser>::EncodedSize,
+                <<CS::Kem as Kem>::DecapsulationKey as KeySizeUser>::KeySize,
+                <<CS::Kem as Kem>::EncapsulationKey as KeySizeUser>::KeySize,
             >,
             Output = <<CS::Pake as Pake>::Output as EncodedSizeUser>::EncodedSize,
         >,
 {
     type EncodedSize = Sum<
         Sum<
-            <<CS::Kem as KemCore>::DecapsulationKey as EncodedSizeUser>::EncodedSize,
-            <<CS::Kem as KemCore>::EncapsulationKey as EncodedSizeUser>::EncodedSize,
+            <<CS::Kem as Kem>::DecapsulationKey as KeySizeUser>::KeySize,
+            <<CS::Kem as Kem>::EncapsulationKey as KeySizeUser>::KeySize,
         >,
         <<CS::Pake as Pake>::Output as EncodedSizeUser>::EncodedSize,
     >;
 
-    fn from_bytes(enc: &Encoded<Self>) -> Self {
+    fn from_bytes(enc: &Encoded<Self>) -> core::result::Result<Self, PakeKemError> {
         let (enc, pake_output) = enc.split_ref();
         let (dk_bytes, ek_bytes) = enc.split_ref();
 
-        Self {
-            pake_output: <CS::Pake as Pake>::Output::from_bytes(pake_output),
-            dk: <CS::Kem as KemCore>::DecapsulationKey::from_bytes(dk_bytes),
-            ek: <CS::Kem as KemCore>::EncapsulationKey::from_bytes(ek_bytes),
-        }
+        Ok(Self {
+            pake_output: <CS::Pake as Pake>::Output::from_bytes(pake_output)?,
+            dk: <<CS::Kem as Kem>::DecapsulationKey as KeyInit>::new(dk_bytes),
+            ek: <<CS::Kem as Kem>::EncapsulationKey as TryKeyInit>::new(ek_bytes)
+                .map_err(|_| PakeKemError::Deserialization)?,
+        })
     }
 
     fn as_bytes(&self) -> Encoded<Self> {
-        self.dk
-            .as_bytes()
-            .concat(self.ek.as_bytes())
+        KeyExport::to_bytes(&self.dk)
+            .concat(KeyExport::to_bytes(&self.ek))
             .concat(self.pake_output.as_bytes())
     }
 }